requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
anyio==3.6.2
    # via
    #   httpcore
    #   starlette
    #   watchfiles
bcrypt==4.0.1
    # via passlib
beanie==1.19.1
    # via -r requirements.in
black==23.3.0
    # via -r requirements.in
build==0.10.0
    # via pip-tools
certifi==2023.5.7
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.15.1
    # via cryptography
charset-normalizer==3.1.0
    # via requests
click==8.1.3
    # via
    #   beanie
    #   black
    #   pip-tools
    #   uvicorn
cryptography==40.0.2
    # via python-jose
dnspython==2.3.0
    # via
    #   -r requirements.in
    #   email-validator
    #   pymongo
ecdsa==0.18.0
    # via python-jose
email-validator==2.0.0.post2
    # via pydantic
exceptiongroup==1.1.1
    # via pytest
fastapi==0.95.2
    # via -r requirements.in
flake8==6.0.0
    # via -r requirements.in
gunicorn==20.1.0
    # via -r requirements.in
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==0.17.2
    # via httpx
httptools==0.5.0
    # via uvicorn
httpx==0.24.1
    # via -r requirements.in
idna==3.4
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
iniconfig==2.0.0
    # via pytest
lazy-model==0.0.5
    # via beanie
mccabe==0.7.0
    # via flake8
motor==3.1.2
    # via
    #   -r requirements.in
    #   beanie
mypy-extensions==1.0.0
    # via black
packaging==23.1
    # via
    #   black
    #   build
    #   pytest
passlib[bcrypt]==1.7.4
    # via -r requirements.in
pathspec==0.11.1
    # via black
pip-tools==6.13.0
    # via -r requirements.in
platformdirs==3.5.1
    # via black
pluggy==1.0.0
    # via pytest
pyasn1==0.5.0
    # via
    #   python-jose
    #   rsa
pycodestyle==2.10.0
    # via flake8
pycparser==2.21
    # via cffi
pydantic[email]==1.10.7
    # via
    #   -r requirements.in
    #   beanie
    #   fastapi
    #   lazy-model
pyflakes==3.0.1
    # via flake8
pymongo==4.3.3
    # via motor
pyproject-hooks==1.0.0
    # via build
pytest==7.3.1
    # via
    #   -r requirements.in
    #   pytest-asyncio
    #   pytest-env
    #   pytest-randomly
pytest-asyncio==0.21.0
    # via -r requirements.in
pytest-env==0.8.1
    # via -r requirements.in
pytest-randomly==3.13.0
    # via -r requirements.in
python-dotenv==1.0.0
    # via uvicorn
python-jose[cryptography]==3.3.0
    # via -r requirements.in
python-multipart==0.0.6
    # via -r requirements.in
pyyaml==6.0
    # via uvicorn
requests==2.31.0
    # via -r requirements.in
rsa==4.9
    # via python-jose
six==1.16.0
    # via ecdsa
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
starlette==0.27.0
    # via fastapi
toml==0.10.2
    # via beanie
tomli==2.0.1
    # via
    #   black
    #   build
    #   pyproject-hooks
    #   pytest
typing-extensions==4.6.0
    # via pydantic
urllib3==2.0.2
    # via requests
uvicorn[standard]==0.22.0
    # via -r requirements.in
uvloop==0.17.0
    # via uvicorn
watchfiles==0.19.0
    # via uvicorn
websockets==11.0.3
    # via uvicorn
wheel==0.40.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
